% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.r
\name{createConfig}
\alias{createConfig}
\title{Creates a configuration with connection parameters}
\usage{
createConfig(
  pghost = NULL,
  pgport = NULL,
  pgdatabase = NULL,
  pguser = NULL,
  pgpassword = NULL,
  lagoon_host = "localhost",
  lagoon_port = "22089",
  ingest_user = unname(Sys.info()["user"])
)
}
\arguments{
\item{pghost}{The host of the PostgreSQL database.}

\item{pgport}{The port of the database.}

\item{pgdatabase}{The name of the database.}

\item{pguser}{The name of the database user.}

\item{pgpassword}{The password of the database user.}

\item{lagoon_host}{The host of the Lagoon server.
Defaults to "localhost" or the value of the environment
variable LAGOON_HOST if defined.}

\item{lagoon_port}{The port of the Lagoon server.
Defaults to "22089" or the value of the environment
variable LAGOON_HOST if defined.}

\item{ingest_user}{The user of ingest. Defaults to the system user.}
}
\description{
Creates a configuration with connection parameters
}
