% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.r
\name{getDefaultConfig}
\alias{getDefaultConfig}
\title{The default configuration to use for calls in this package.}
\usage{
getDefaultConfig()
}
\description{
The default configuration to use for calls in this package.
}
