% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.r
\name{getTableNames}
\alias{getTableNames}
\title{Lists the table names along with source names}
\usage{
getTableNames(config = getDefaultConfig())
}
\description{
Lists the table names along with source names
}

