% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.r
\name{ingest}
\alias{ingest}
\title{Imports data into a PostgreSQL database}
\usage{
ingest(source_name, source_path, description = NULL, tags = NULL,
  json_path = NULL, config = getDefaultConfig())
}
\arguments{
\item{source_name}{The name to give to the source.}

\item{source_path}{The location of the data to import. Can be a local file
or a url.}

\item{description}{Free-form description for this data source. If not
provided, the source name is used instead.}

\item{tags}{A character vector of tags to apply to the source.}

\item{json_path}{Ingest part of a JSON file. A path is either '_',
'[path]' or '{"field":path}'. For example, '{"groups":[_]}'
extracts all elements of the array of the "groups" field of
the top-level object as separate rows into the database.}

\item{config}{specifies the database connection details. Defaults to
defaultConfiguration.}
}
\description{
Imports data into a PostgreSQL database
}

