% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrlagoon.r
\name{sql_translate_env.LagoonConnection}
\alias{sql_translate_env.LagoonConnection}
\title{Translates expressions from R to SQL containing json operators.}
\usage{
\method{sql_translate_env}{LagoonConnection}(con)
}
\description{
The infix JSON operators are surrounded with \% in R (\%->\% translates to ->
in SQL, \%@>\% translates to @>, etc). We translate all the operators described
in https://www.postgresql.org/docs/9.5/static/functions-json.html
}
\details{
jsonb('{...}') is translated to PostgreSQL syntax '{...}'::jsonb.
}
\examples{
\dontrun{
 # Test with
 dbConn <- src_lagoon("localhost", "3000")
 translate_sql( jsonb('{"a":2}')\%->\%'a', con=dbConn[[1]])
}
}
