% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrlagoon.r
\name{src_lagoon}
\alias{src_lagoon}
\title{Call this function to connect to a Lagoon server.}
\usage{
src_lagoon(
  host = "localhost",
  port = "22089",
  user = unname(Sys.info()["user"]),
  prot = "http",
  ...
)
}
\description{
See the documentation of sql_translate_env.LagoonConnection to learn
how JSON operators are translated from R.
}
\examples{
\dontrun{
require(dplyr); require(RLagoon)
db <- src_lagoon(host = "localhost", port = "3000", user = "mary")
tbl(db, "someTable")
}
}
